refactor(node): move node utilities to adapter
