+++ .changeset/move-node-utils.md
@@ -1,0 +1,5 @@
+---
+'@lapidist/design-lint': patch
+---
+
+move node-specific utilities to adapter directory
