+++ src/formatters/stylish.ts
@@ -1,5 +1,5 @@
 import type { LintResult } from '../core/types.js';
-import { relFromCwd } from '../utils/paths.js';
+import { relFromCwd } from '../adapters/node/utils/paths.js';
 
 const codes = {
   red: (s: string) => `\x1b[31m${s}\x1b[0m`,
