+++ tests/core/glob-ignore.test.ts
@@ -3,7 +3,7 @@ import assert from 'node:assert/strict';
 import fs from 'node:fs';
 import path from 'node:path';
 import { FileSource } from '../../src/adapters/node/file-source.ts';
-import { makeTmpDir } from '../../src/utils/tmp.ts';
+import { makeTmpDir } from '../../src/adapters/node/utils/tmp.ts';
 import type { Config } from '../../src/core/linter.ts';
 
 void test('FileSource.scan applies nested ignore files for glob targets', async () => {
