+++ tests/tmp-utils.test.ts
@@ -1,7 +1,7 @@
 import test from 'node:test';
 import assert from 'node:assert/strict';
 import fs from 'node:fs';
-import { makeTmpDir } from '../src/utils/tmp.ts';
+import { makeTmpDir } from '../src/adapters/node/utils/tmp.ts';
 
 void test('makeTmpDir creates a temporary directory', () => {
   const dir = makeTmpDir('test-');
